Move lambdas to separate package
