+++ database/test/test_mml_loader.py
@@ -3,9 +3,10 @@ from typing import Callable
 
 import codes
 import pytest
-from mml_loader.mml_loader import MMLLoader
 from requests_mock.request import _RequestObjectProxy
 from sqlalchemy.orm import Session
+
+from lambdas.mml_loader.mml_loader import MMLLoader
 
 
 @pytest.fixture()
