+++ database/test/test_ryhti_client.py
@@ -8,9 +8,10 @@ import codes
 import models
 import pytest
 from requests_mock.request import _RequestObjectProxy
-from ryhti_client.ryhti_client import RyhtiClient
 from simplejson import JSONEncoder
 from sqlalchemy.orm import Session
+
+from lambdas.ryhti_client.ryhti_client import RyhtiClient
 
 from .conftest import deepcompare
