+++ docker-compose.dev.yml
@@ -12,7 +12,7 @@ services:
   db_manager:
     build:
       context: .
-      dockerfile: database/db_manager.Dockerfile
+      dockerfile: lambdas/db_manager/db_manager.Dockerfile
     env_file:
       - .env
     environment:
@@ -24,12 +24,12 @@ services:
     depends_on:
       - db
     volumes:
-      - ./database/db_manager/db_manager.py:/var/task/app.py
+      - ./lambdas/db_manager/db_manager.py:/var/task/app.py
 
   koodistot_loader:
     build:
       context: .
-      dockerfile: database/koodistot_loader.Dockerfile
+      dockerfile: lambdas/koodistot_loader/koodistot_loader.Dockerfile
     env_file:
       - .env
     environment:
@@ -41,12 +41,12 @@ services:
     depends_on:
       - db
     volumes:
-      - ./database/koodistot_loader/koodistot_loader.py:/var/task/app.py
+      - ./lambdas/koodistot_loader/koodistot_loader.py:/var/task/app.py
 
   ryhti_client:
     build:
       context: .
-      dockerfile: database/ryhti_client.Dockerfile
+      dockerfile: lambdas/ryhti_client/ryhti_client.Dockerfile
     env_file:
       - .env
     environment:
@@ -67,13 +67,13 @@ services:
       - db
       - mock_xroad
     volumes:
-      - ./database/ryhti_client/ryhti_client.py:/var/task/app.py
-      - ./database/ryhti_debug:/var/task/ryhti_debug
+      - ./lambdas/ryhti_client/ryhti_client.py:/var/task/app.py
+      - ./ryhti_debug:/var/task/ryhti_debug
 
   mml_loader:
     build:
       context: .
-      dockerfile: database/mml_loader.Dockerfile
+      dockerfile: lambdas/mml_loader/mml_loader.Dockerfile
     env_file:
       - .env
     environment:
@@ -86,7 +86,7 @@ services:
     depends_on:
       - db
     volumes:
-      - ./database/mml_loader/mml_loader.py:/var/task/app.py
+      - ./lambdas/mml_loader/mml_loader.py:/var/task/app.py
 
   mock_xroad:
       image: mockserver/mockserver:latest
